+++ pyobs/interfaces/IDataSequence.py
@@ -38,7 +38,7 @@ class IDataSequence(IAbortable, metaclass=ABCMeta):
                 during the wait.
 
         Raises:
-            GrabImageError: If the device is already busy (exposing or already running a
+            DeviceBusyError: If the device is already busy (exposing or already running a
                 sequence).
         """
         ...