+++ pyobs/interfaces/IRunnable.py
@@ -11,7 +11,13 @@ class IRunnable(IAbortable, metaclass=ABCMeta):
 
     @abstractmethod
     async def run(self, **kwargs: Any) -> None:
-        """Perform module task"""
+        """Perform module task
+
+        Raises:
+            ValueError: If this task is already running.
+            ScriptError: ScriptRunner-based implementations wrap whatever the underlying script
+                raises that isn't already a domain exception.
+        """
         ...